fix: fix test error of group-mark
